6 hours
Removed sort order functionality temporarily. Not sure whether we need it or not.

Beginning the process of adding and manipulating templates.

Should have merged back to master a commit or two ago, but will now.  Not sure if we maintain JNBPA as client so slightly pivoting and prioritizing Bytion/CS/WFL development until a commitment is reached.