6 hours
Fixed extraneous logging issue. Issue stemmed from the fact that I wasn’t a flag and re-instantiating the .on listeners and not closing them except for BB close(). Added the appropriate flags.  Going to tackle META issues. More than likely related to activation due to the removal of activate callers in place of sys bus subscriptions.